configuration parameters fix

+++ Projects/FSAgent/FSAgentServer/PatchManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.IO;
+using Infrastructure.Common;
 
 namespace FSAgentServer
 {
@@ -10,6 +11,19 @@ namespace FSAgentServer
 	{
 		public static void Patch()
 		{
+			try
+			{
+				Patch1();
+			}
+			catch { }
+		}
+
+		static void Patch1()
+		{
+			var patchNo = PatchHelper.GetPatchNo("FSAgent");
+			if (patchNo > 0)
+				return;
+
 			if (Directory.Exists("Pictures"))
 			{
 				Directory.Delete("Pictures", true);
@@ -18,6 +32,8 @@ namespace FSAgentServer
 			{
 				Directory.Delete("Logs", true);
 			}
+
+			PatchHelper.SetPatchNo("FSAgent", 1);
 		}
 	}
 }
